requirements.txt
pyyaml
sseclient-py
